Refactor clustering: introduce AbstractCorpus and Subcorpus classes.
