+++ tempo_embeddings/text/subcorpus.py
@@ -1,0 +1,44 @@
+from typing import TYPE_CHECKING
+from typing import Optional
+from numpy.typing import ArrayLike
+from .abstractcorpus import AbstractCorpus
+from .passage import Passage
+
+
+if TYPE_CHECKING:
+    from .corpus import Corpus
+
+
+class Subcorpus(AbstractCorpus):
+    """A partial corpus referencing to a parent corpus."""
+
+    def __init__(
+        self, parent_corpus: "Corpus", indices: list[int], label: Optional[str] = None
+    ) -> None:
+        self._corpus = parent_corpus
+        self._indices = indices
+        self._label = label
+
+    def __repr__(self) -> str:
+        return f"Subcorpus({self._label!r}, {self._indices[:10]!r})"
+
+    def __add__(self, other: "Subcorpus", new_label: Optional[str] = None):
+        if self._corpus == other._corpus:
+            indices = set(self._indices + other._indices)
+            label = new_label or "+".join([self._label, other._label])
+        else:
+            raise ValueError("Cannot merge sub-corpora with different parent corpora.")
+
+        return Subcorpus(self._corpus, list(indices), label)
+
+    @property
+    def passages(self) -> list[Passage]:
+        return [self._corpus.passages[i] for i in self._indices]
+
+    @property
+    def embeddings(self) -> ArrayLike:
+        return (
+            None
+            if self._corpus.embeddings is None
+            else self._corpus.embeddings.take(self._indices, axis=0)
+        )
