+++ tempo_embeddings/visualization/wizmap.py
@@ -10,7 +10,7 @@ from http.server import SimpleHTTPRequestHandler
 from typing import Optional
 import numpy as np
 import wizmap
-from ..text.corpus import Corpus
+from ..text.abstractcorpus import AbstractCorpus
 from .visualizer import Visualizer
 
 
@@ -19,7 +19,11 @@ class WizmapVisualizer(Visualizer):
     grid_file_name = "grid.json"
 
     def __init__(
-        self, corpus: Corpus, title: str, path: str = None, stopwords: list[str] = None
+        self,
+        corpus: AbstractCorpus,
+        title: str,
+        path: str = None,
+        stopwords: list[str] = None,
     ):
         super().__init__(corpus)
 
@@ -72,7 +76,7 @@ class WizmapVisualizer(Visualizer):
     def _write_data(self):
         """Write Wizmap visualizations to a file."""
 
-        embeddings = np.array(self._corpus.umap_embeddings())
+        embeddings = np.array(self._corpus.embeddings)
         xs = embeddings[:, 0].astype(float).tolist()
         ys = embeddings[:, 1].astype(float).tolist()
