+++ tests/unit/text/test_corpus.py
@@ -21,79 +21,6 @@ class TestCorpus:
         )
 
     @pytest.mark.parametrize(
-        "corpus,token,metadata,expected",
-        [
-            (Corpus(), "test", {}, Corpus()),
-            (
-                Corpus([Passage("test line")]),
-                "test",
-                {},
-                Corpus([Passage("test line", highlighting=Highlighting(0, 4))]),
-            ),
-            (
-                Corpus(
-                    [Passage("test line1"), Passage("test line2")],
-                ),
-                "line1",
-                {},
-                Corpus([Passage("test line1", highlighting=Highlighting(5, 10))]),
-            ),
-            (
-                Corpus(
-                    [
-                        Passage("test line1", {"test": "value1"}),
-                        Passage("test line2", {"test": "value2"}),
-                    ],
-                ),
-                "test",
-                {"test": "value1"},
-                Corpus(
-                    [
-                        Passage(
-                            "test line1",
-                            {"test": "value1"},
-                            highlighting=Highlighting(0, 4),
-                        )
-                    ]
-                ),
-            ),
-            (
-                Corpus(
-                    [
-                        Passage("test line1", {"key1": "value1", "key2": "value2"}),
-                        Passage("test line2", {"key1": "value1", "key2": "value3"}),
-                    ]
-                ),
-                "test",
-                {"key1": "value1", "key2": "value2"},
-                Corpus(
-                    [
-                        Passage(
-                            "test line1",
-                            {"key1": "value1", "key2": "value2"},
-                            highlighting=Highlighting(0, 4),
-                        )
-                    ],
-                ),
-            ),
-            (
-                Corpus(
-                    [
-                        Passage("test line1", {"test": "value1"}),
-                        Passage("test line2", {"test": "value2"}),
-                    ]
-                ),
-                "test",
-                {"test": "value3"},
-                Corpus(),
-            ),
-        ],
-    )
-    def test_subcorpus(self, corpus, token, metadata, expected):
-        # TODO: test with exact_match=True
-        assert corpus.subcorpus(token, exact_match=False, **metadata) == expected
-
-    @pytest.mark.parametrize(
         "corpus,metadata_fields,expected",
         [
             (Corpus(), None, []),
